partial menu class updated
