+++ RestaurantMenu/Model/Core/MenuDir/Menu.cs
@@ -7,7 +7,7 @@ using Model.Core.MealDir;
 
 namespace Model.Core.MenuDir
 {
-    public abstract class Menu : IMenu
+    public abstract partial class Menu : IMenu
     {
         public Menu() { }
 
@@ -27,25 +27,6 @@ namespace Model.Core.MenuDir
             return Meals;
         }
         public void CloseMenu() { }
-
-        public void DeleteMeal(Meal meal) {
-            _meals = _meals.Where(m => m != meal).ToArray();
-        }
-
-        public static Menu operator +(Menu menu, Meal meal)
-        {
-            Meal[] updatedMeals = new Meal[menu._meals.Length + 1];
-            Array.Copy(menu._meals, updatedMeals, menu._meals.Length);
-            updatedMeals[updatedMeals.Length - 1] = meal;
-
-            menu._meals = updatedMeals;
-            return menu; 
-        }
-
-        public void AddMeal(Meal meal)
-        {
-            var _ = this + meal;
-        }
 
     }
 }
